refactor(layout): move appContext to src/runtime/ for CLI+MCP sharing
