+++ src/commands/add/__tests__/add.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { add } from '../index.js';
-import type { AppContext } from '../../../cli/appContext.js';
+import type { AppContext } from '../../../runtime/appContext.js';
 import type { Company } from '../../../utils/types/company.js';
 
 // Build a fake AppContext where every repository/service method is a vi.fn() spy.
