+++ src/mcp/tools.ts
@@ -196,7 +196,7 @@ Returns what's missing and what the next step should be.`,
     async () => {
       try {
         const { loadConfig } = await import('../utils/config.js');
-        const { createAppContext } = await import('../cli/appContext.js');
+        const { createAppContext } = await import('../runtime/appContext.js');
         // loadConfig validates wolf.toml exists; createAppContext opens SQLite + loads profile.
         await loadConfig();
         const ctx = createAppContext();
